+++ crates/node_binding/binding.d.ts
@@ -160,12 +160,18 @@ export interface RawEntryItem {
 export interface RawExperiments {
   lazyCompilation: boolean
   incrementalRebuild: boolean
+  asyncWebAssembly: boolean
 }
 export interface RawExternalItem {
   type: "string" | "regexp" | "object"
   stringPayload?: string
   regexpPayload?: string
-  objectPayload?: Record<string, string>
+  objectPayload?: Record<string, RawExternalItemValue>
+}
+export interface RawExternalItemValue {
+  type: "string" | "bool"
+  stringPayload?: string
+  boolPayload?: boolean
 }
 export interface RawExternalsPresets {
   node: boolean
@@ -298,6 +304,7 @@ export interface RawOutputOptions {
   path: string
   publicPath: string
   assetModuleFilename: string
+  webassemblyModuleFilename: string
   filename: string
   chunkFilename: string
   cssFilename: string
@@ -308,6 +315,8 @@ export interface RawOutputOptions {
   enabledLibraryTypes?: Array<string>
   globalObject: string
   importFunctionName: string
+  iife: boolean
+  module: boolean
 }
 export interface RawResolveOptions {
   preferRelative?: boolean
@@ -426,6 +435,7 @@ export interface JsHooks {
   afterEmit: (...args: any[]) => any
   make: (...args: any[]) => any
   optimizeChunkModule: (...args: any[]) => any
+  finishModules: (...args: any[]) => any
   normalModuleFactoryResolveForScheme: (...args: any[]) => any
 }
 export interface JsModule {