+++ crates/rspack_core/src/stats.rs
@@ -126,7 +126,6 @@ impl Stats<'_> {
       .compilation
       .chunk_by_ukey
       .values()
-      .into_iter()
       .map(|c| -> Result<_> {
         let mut files = Vec::from_iter(c.files.iter().cloned());
         files.sort_unstable();