+++ packages/rspack/src/config/normalization.ts
@@ -52,12 +52,15 @@ export const getNormalizedRspackOptions = (
 				cssFilename: output.cssFilename,
 				cssChunkFilename: output.cssChunkFilename,
 				assetModuleFilename: output.assetModuleFilename,
+				webassemblyModuleFilename: output.webassemblyModuleFilename,
 				uniqueName: output.uniqueName,
 				enabledLibraryTypes: output.enabledLibraryTypes
 					? [...output.enabledLibraryTypes]
 					: ["..."],
 				globalObject: output.globalObject,
 				importFunctionName: output.importFunctionName,
+				iife: output.iife,
+				module: output.module,
 				library: libraryBase && {
 					type:
 						output.libraryTarget !== undefined