+++ packages/rspack/tests/configCases/externals/array-externals/inject.js
@@ -1,3 +1,2 @@
 global.foo = "foo";
-global.bar = "bar";
-global.react = "react";
+global.raz = "raz";