+++ npm/darwin-arm64/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @rspack/binding-darwin-arm64
 
+## 0.1.2
+
 ## 0.1.1
 
 ### Patch Changes