+++ packages/create-rspack/CHANGELOG.md
@@ -1,5 +1,14 @@
 # create-rspack
 
+## 0.1.2
+
+### Patch Changes
+
+- 6d482a7: fix yargs default command ad help
+- 8f8e025: align optimizeChunkModules
+- 0f0969a: feat: add the handling of duplicate project names
+- 8c196ee: feat: support package manager detect
+
 ## 0.1.1
 
 ### Patch Changes